requirements.txt
streamlit>=1.25
pandas>=2.0
numpy>=1.25
scikit-learn>=1.3
matplotlib>=3.8
seaborn>=0.12
xgboost>=1.7
google-generativeai>=0.1.0
python-dotenv>=1.0
